Add parameter to control output of RVA+offset
Add a new param `includeModOffset` which when true, and if private
symbols are used, will cause the output to be suffixed with
(module+RVA) details.

This is useful in some very niche cases where we have a frame, but need
to get back to its RVA - a "reverse lookup" of sorts.